Gizmo adding
The square can now be added to the board through the GUI in build mode,
however an exception is thrown when a second is added because they both
have the same default position. We can discuss what to do in this
situation, for now Tom has it throwing an error. I think we would be
better of placing the gizmo on the next available position instead of
throwing an exception (it's not exactly user friendly).