Updated the Model.
Fixed the bug where it was possible to place gizmos on top of already placed gizmos.

Altered the Model to ensure removeGizmo(g) removes al the refs to the gizmo object held in the model.

Altered the Grid to make it possible to retrieve the gizmo placed at a grid position. added the getGizmo(int x, int y) method to return the gizmo at the given cord.

Updated Model.Global with gravity and friction. Further updated the model to enable these to be set.

Created Model.IBall interface for Model.Ball

Removed the package model.physics as per discussion. This means that the jar (libs/gizmo.jar) is now a required component of the build. The build path should be updated to relfect thsi.

Added package-info documentation to all of the packages.