Worked on file loading.
Can now load and move balls.

Users can now select the file they wish to save to/load from using a filechooser. This was done with a wrapper around the JFileChooser to make useing it easy from the controler. This code was taken from my foliotracker application.

Modified the update in BuildBoard so that in the case of no args all of the gizmos registered with the model are retreived to be drawn in the future.