GUI, Model and other Changes
Updated the Board Views, they now have an abstract superclass containing
similar methods and attributes.

Added the functionality to rotate any Gizmo located on the screen. I
also changed the model to make these changes to the objects.

Gizmos can now be deleted. The removeGizmo method by tom works, I only
had to pass this to the view to delete the last reference to the object.

Some other minor changes and bug fixes.

Everything seems to be working.