Fixed multiple inconsistencies in how model and view are handling gizmo coordinates.
In runBoard.paintComponent() the width of the circle being painted was retrieved from ball.getRadius(). However this value was not being multiplied by 2 (width=radius*2), resulting in a ball of half the size of the ball in model being drawn.

In addition to this, the radius of the Ball in model was set to 1L as default. This would create a ball 2L wide, rather than 1L wide. These two bugs combined to cause a ball 1L wide being drawn on the screen but a ball of 2L wide being used for collision calculations in model.

In paint component, when calculating the x,y coordinates to draw the ball, the radius returned from the model was not being converted from L to pixels.

In drawGizmos(), the x & y coordinates were being retrieved from gizmo.getX() and gizmo.getY(). However, in the model x and y points to the center of the gizmo. However when drawing gizmos, the Graphics2D class treats the x and y cooridinates as the top left point to be painted, resulting in gizmos being painted in the wrong positition (slightly). This would make it look like the ball was coliding at the wrong time.

in timeUntilCollision() if a gizmo is a circle, it's radius is calulated using circle width/2, rather than hard coded 0.5