Added logging
Added loging using the standard java logging to tidy the console output we are generating.

Updated the board to remove duplicate code.

Updated move gizmo in the board so it works for gizmos larger than 1L x 1L and so it catched invalid positions.

Redone the javaDoc in the board, so its up to date.