Collisions Update
I've added the corner collisions to the squares and triangles. The
collisions look pretty solid now. I am more than happy to submit this as
the final prototype.

NOTE: The TimeUntilCollision method is getting pretty lengthy and I plan
on breaking it down to make it more modular and easily maintainable.