Added update to Gizmos
This means that the gizmos states can now be updated.

The updated methods calculates a percentage that represents the current progress through the current trigger, and then calls action() (see below).

Added an action method, that is repsonible for performing the action of the gizmo based in the trigger time information (triggerPercentage, and time values) avialible to it.