Model now supports multiple balls.
Updated the model so that it now supports multiple balls during one game. This includes collisions between the balls.

Added a new file ballCollisions.txt to demo this feature.

Unfortunatly I am not sure how to best clear the balls stored in the view. As such it is not removing balls there, such as on load. This is not an issue in the model