FLIPPER COLLISIONS! YAAAY!! But......
Now the flipper collisions are working, so that's good.

2 BUGS:

The first one is a continuation off Cameron's bug (cheers mate) that
causes the flipper to now trigger itself when hit.

The second bug is something we should discuss, the collision detected
works perfectly when the ball is moving TOWARDS the flipper. It even
works at any angle you can throw at it. However if the ball is moving
away from the flipper and the flipper is triggered and manages to hit
the ball then the ball gets stuck inside the flipper! So yeah that's a
minorish bug to fix.