Ball placement and memory leak fix
Balls can no longer be added on top of already placed gizmos. See issue #30

Fix for issue #37. All connections are now recorded in the model allows the deletion of the gizmos to remove all connections to ensure to refs are kept. During the fixing of this I have added methods in the model (model.IModel) and in the gizmos to handle the disconnecting of the gizmos. This is part of issue #36 and #13