Flippers now collapse after trigger from ball/gizmo
Modified the trigging system to use a enum to represent the type of the trigger. This means its possible to work out what action we wish to complete depending on the trigger type. At the moment this is only used in the flippers to handle ball and gizmo triggers so that the flipper collapses again.

This could also possibly effect issue #35.