User setting the ball velocity
Ball now added with default velocity of 0, 50.

Modified the add absorber to comply with the spec of the Model.IModel interface, removing added method in Model

Add a set ball velocity window, users can now set the velocity of the ball when they wish.

Users can now specify the velocity that the ball leaves the absorber at.

Updated javaDoc throughout and added commentary to the controller package.