Fixed issue with OpenJDK crashing everything
Added a check to ensure that gizmo x/y positions are greater then 0 and that the width/height values are greater than 0.

Removed a whole bunch of system outs.

Handling OpenJDK's terrible implementation of the JVM and alerting the user that some things wont work, and as such are turned off.